+++ test/pay/stripe/subscription_test.rb
@@ -317,7 +317,10 @@ class Pay::Stripe::SubscriptionTest < ActiveSupport::TestCase
       current_period_start: 1486568724,
       customer: "cus_1234",
       ended_at: nil,
-      latest_invoice: "in_1000",
+      latest_invoice: {
+        id: "in_1000",
+        status: "paid"
+      },
       plan: {
         id: "default"
       },